+++ src/badguy/treewillowisp.cpp
@@ -127,9 +127,9 @@ TreeWillOWisp::active_update(float elapsed_time)
   }
 
   angle = fmodf(angle + elapsed_time * speed, static_cast<float>(2*M_PI));
-  Vector newpos(start_position + Vector(sin(angle) * radius, 0));
+  Vector newpos(start_position + Vector(sinf(angle) * radius, 0));
   movement = newpos - get_pos();
-  float sizemod = cos(angle) * 0.8f;
+  float sizemod = cosf(angle) * 0.8f;
   /* TODO: modify sprite size */
 
   sound_source->set_position(get_pos());
