+++ src/object/moving_sprite.cpp
@@ -199,8 +199,8 @@ void MovingSprite::spawn_explosion_sprites(int count, const std::string& sprite_
       Vector ppos = bbox.get_middle();
       float angle = graphicsRandom.randf(static_cast<float>(-M_PI_2), static_cast<float>(M_PI_2));
       float velocity = graphicsRandom.randf(350, 400);
-      float vx = sin(angle)*velocity;
-      float vy = -cos(angle)*velocity;
+      float vx = sinf(angle)*velocity;
+      float vy = -cosf(angle)*velocity;
       Vector pspeed = Vector(vx, vy);
       Vector paccel = Vector(0, Sector::current()->get_gravity()*10);
       Sector::current()->add_object(std::make_shared<SpriteParticle>(sprite_path,
